requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
pydantic==2.8.2
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pika==1.3.2
requests==2.32.3
google-generativeai>=0.6.0
passlib==1.7.4
PyJWT==2.9.0
